Added COUNT branch
Added a branch to increment the value of R7, as adding the result
of R0 (value) AND R1 (mask) doesn't work, as the reslt of the AND
is not #1, but rather some number with either 1 or 0 bits set.

Thus, if the result of the AND is positive, then we want to branch
out of the normal loop, increment R7, and branch back to the top
of the main loop.